feat: Apply spring-data-jpa:Intake

+++ src/main/java/com/safefood/controller/IntakeController.java
@@ -83,7 +83,7 @@ public class IntakeController {
     }
 
     @GetMapping("/delete")
-    private String intakeDelete(int intakeCode) {
+    private String intakeDelete(Long intakeCode) {
         intakeService.deleteIntake(intakeCode);
         log.info(String.valueOf(intakeCode));
